Spike gadgets edits, edits for Phy2 dev

+++ kilosort2-pipeline/functions/importChannels.m
@@ -48,10 +48,10 @@ for i = 1:length(channels)
     junk = fread(fid,channels(i)-1,'int16'); %skip ahead to the channel
     channelData = fread(fid,[1,inf],'1*int16=>int16',(2*headerSize)+2+(NumChannels*2))';
     frewind(fid);
-       
+        
     %commeted out bc SpikeGadgets output is already inverted 
 %     channelData = double(channelData)*-1; %reverse the sign to make spike point up
-    channelData = channelData * 12780; %convert to uV (for Intan digital chips)
+    channelData = double(channelData) * 12780; %convert to uV (for Intan digital chips)
     channelData = channelData / 65536;
     recData = [recData channelData];
